Email plugin improvements, including adding grmt file to email, and bug fixes, including queuing up attachments when I want to send multiple emails. Also initial .webp support, and playing around with picture sizes
