+++ src/gourmand/plugin.py
@@ -412,6 +412,7 @@ class PluginPlugin (StandardPlugin):
 class RecDisplayModule (UIModule):
 
     def __init__ (self, recDisplay):
+        self.action_groups = [] #wv
         self.rd = recDisplay; self.rg = self.rd.rg
         UIModule.__init__(self)
